requirements.txt
openai
requests
python-dotenv
chatterbox-tts
pydub
ffmpeg-python
mlx-whisper
tqdm
gradio
